requirements.txt
asttokens==2.4.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
clarabel==0.6.0
comm==0.1.4
contourpy==1.2.0
coverage==7.3.2
cvxpy==1.4.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
easydict==1.11
ecos==2.0.12
exceptiongroup==1.1.3
executing==2.0.0
faiss-cpu==1.7.4
faiss-gpu==1.7.2
filelock==3.12.4
flake8==3.9.2
fonttools==4.46.0
fsspec==2023.9.2
gdown==4.7.1
h5py==3.10.0
idna==3.4
iniconfig==2.0.0
invoke==1.7.3
ipykernel==6.25.2
ipython==7.34.0
isort==4.3.21
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
kornia==0.7.0
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
osqp==0.6.3
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prettytable==0.7.2
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-term==0.6
pybind11==2.11.1
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.3
pytest-cov==2.12.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
qdldl==0.1.7.post0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
scs==3.2.3
seaborn==0.13.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
torch==2.1.0
torchmetrics==1.2.1
torchnet==0.0.4
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
visdom==0.2.4
wcwidth==0.2.8
websocket-client==1.6.4
